실행 cpp 파일 내의 헤더파일로 분리할 수 있는 부분 분리
